// src/lib/api/openapi.ts
export interface paths {
    "/rest/comics": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List comic. */
        get: operations["listComic"];
        put?: never;
        /** Add comic. */
        post: operations["addComic"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/rest/comics/{comicCode}/titles": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List comic title. */
        get: operations["listComicTitle"];
        put?: never;
        /** Add comic title. */
        post: operations["addComicTitle"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/rest/comics/{comicCode}/covers": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List comic cover. */
        get: operations["listComicCover"];
        put?: never;
        /** Add comic cover. */
        post: operations["addComicCover"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}